requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
black==20.8b1
click==7.1.2
Django==2.1.7
djongo==1.3.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
mongoengine==0.23.0
mypy-extensions==0.4.3
pathspec==0.8.1
promise==2.3
pymongo==3.11.3
pytz==2021.1
regex==2021.4.4
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
